+++ index.js
@@ -1,17 +1,26 @@
 "use strict";
 
 const pkg = require("./package.json");
+
+const os = require("os");
 const path = require("path");
 const { createWriteStream } = require("fs");
+
 const makeDir = require("make-dir");
 const archiver = require("archiver");
 const globby = require("globby");
 const nanomatch = require("nanomatch");
+const pLimit = require("p-limit");
 const { findProdInstalls } = require("inspectdep");
 
 const SLS_TMP_DIR = ".serverless";
 const PLUGIN_NAME = pkg.name;
 const IS_WIN = process.platform === "win32";
+
+// Our builds are CPU and I/O intensive. Add some limits.
+// Use number of cpus as concurrency.
+const concurrency = os.cpus().length;
+const limit = pLimit(concurrency);
 
 // Simple, stable union.
 const union = (arr1, arr2) => {
@@ -357,12 +366,14 @@ class Jetpack {
     const bundleName = path.join(SLS_TMP_DIR, `${functionName}.zip`);
 
     // Package.
-    this._log(`Packaging function: ${bundleName}`);
+    this._log(`Start packaging function: ${bundleName}`);
     await this.globAndZip({ bundleName, functionObject });
 
     // Mutate serverless configuration to use our artifacts.
     functionObject.package = functionObject.package || {};
     functionObject.package.artifact = bundleName;
+
+    this._log(`Finish packaging function: ${bundleName}`);
   }
 
   async packageService() {
@@ -374,11 +385,13 @@ class Jetpack {
     const bundleName = path.join(SLS_TMP_DIR, `${serviceName}.zip`);
 
     // Package.
-    this._log(`Packaging service: ${bundleName}`);
+    this._log(`Start packaging service: ${bundleName}`);
     await this.globAndZip({ bundleName });
 
     // Mutate serverless configuration to use our artifacts.
     servicePackage.artifact = bundleName;
+
+    this._log(`Finish packaging service: ${bundleName}`);
   }
 
   async package() {
@@ -407,7 +420,7 @@ class Jetpack {
       .filter((obj) =>
         (servicePackage.individually || obj.individually) && !(obj.disable || obj.artifact)
       )
-      .map((obj) => this.packageFunction(obj))
+      .map((obj) => limit(() => this.packageFunction(obj)))
     );
 
     // We recreate the logic from `packager#packageService` for deciding whether